Move Firebase auth implementation into app target
